mexec.js is modified
codes are added fro 1)making directories for each run. 2) create setting
info for each run, so that R scripts can ready the parameters. 3) call
the runModels() to execute the R scripts.
Todo: directories containing the R scripts and all the users (not a
single user) are not decided yet. these paths are set to 'xxxxxx'
the data inside the request body is not clear yet. The code assume that
req.body has an array of runs, where each element in the array contains
the runId, scriptName, and so on.